[CodeCompletion] Move ImportDepth to its own file
